requirements.txt
streamlit
langchain
langchain-openai
azure-identity
python-dotenv
beautifulsoup4
pypdf2
faiss-cpu
sentence-transformers
langchain-google-genai
langchain-community
pypdf